# api-spec.yaml
openapi: 3.0.3
info:
  title: ROD Marriage License Database API
  description: |
    API for managing marriage license records with OCR capabilities and authentication.
    This system is designed for LAN-only use by ROD employees.
  version: 1.0.0
  contact:
    name: ROD IT Department
    email: it@rod.gov
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://rod-marriage-db.internal
    description: Internal ROD network server
security:
  - JWT: []
  - BasicAuth: []
paths:
  /api/ocr:
    post:
      summary: Extract text and fields from document using OCR
      description: |
        Processes uploaded documents (PDF, images) to extract text and structured field data.
        Returns confidence scores and verification flags for human review.
      tags:
        - OCR
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
                  description: Document file (PDF, JPG, PNG, TIFF)
              required:
                - document
      responses:
        '200':
          description: OCR processing successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  extractedText:
                    type: string
                    description: Raw OCR text output
                    example: "Application No. 37014 for License to Marry..."
                  extractedFields:
                    type: object
                    description: Structured field data with confidence scores
                    properties:
                      name_spouse1:
                        type: object
                        properties:
                          value:
                            type: string
                            example: "Mary Orlowska"
                          confidence:
                            type: number
                            minimum: 0
                            maximum: 1
                            example: 0.85
                      name_spouse2:
                        type: object
                        properties:
                          value:
                            type: string
                            example: "Josef PaezKowski"
                          confidence:
                            type: number
                            minimum: 0
                            maximum: 1
                            example: 0.92
                      marriage_date:
                        type: object
                        properties:
                          value:
                            type: string
                            format: date
                            example: "1890-02-12"
                          confidence:
                            type: number
                            minimum: 0
                            maximum: 1
                            example: 0.78
                      license_number:
                        type: object
                        properties:
                          value:
                            type: string
                            example: "37014"
                          confidence:
                            type: number
                            minimum: 0
                            maximum: 1
                            example: 0.95
                  needsVerification:
                    type: boolean
                    description: True if any field has low confidence or is missing
                    example: true
                  warnings:
                    type: array
                    items:
                      type: string
                    description: OCR processing warnings
                    example: ["PDF text empty; falling back to OCR"]
                  error:
                    type: string
                    nullable: true
                    description: Error message if processing failed
        '400':
          description: Bad request - no file uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error during OCR processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/licenses:
    get:
      summary: Search marriage license records
      description: |
        Search marriage license records by various criteria.
        Returns paginated results with optional filtering.
      tags:
        - Marriage Licenses
      security:
        - JWT: []
      parameters:
        - name: id
          in: query
          description: Specific record ID
          required: false
          schema:
            type: integer
            example: 123
        - name: name
          in: query
          description: Search by spouse name (partial match)
          required: false
          schema:
            type: string
            example: "Smith"
        - name: date_from
          in: query
          description: Start date for date range search
          required: false
          schema:
            type: string
            format: date
            example: "2023-01-01"
        - name: date_to
          in: query
          description: End date for date range search
          required: false
          schema:
            type: string
            format: date
            example: "2023-12-31"
        - name: license_number
          in: query
          description: Search by license number (partial match)
          required: false
          schema:
            type: string
            example: "37014"
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of records per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarriageLicense'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create new marriage license record
      description: |
        Add a new marriage license record with document upload.
        Can be used with OCR-extracted data or manual entry.
      tags:
        - Marriage Licenses
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name_spouse1:
                  type: string
                  description: First spouse name
                  example: "Mary Orlowska"
                name_spouse2:
                  type: string
                  description: Second spouse name
                  example: "Josef PaezKowski"
                marriage_date:
                  type: string
                  format: date
                  description: Marriage date
                  example: "1890-02-12"
                license_number:
                  type: string
                  description: License or application number
                  example: "37014"
                document:
                  type: string
                  format: binary
                  description: Document file (PDF, JPG, PNG, TIFF)
              required:
                - name_spouse1
                - name_spouse2
                - marriage_date
                - document
      responses:
        '201':
          description: Marriage license record created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    description: New record ID
                    example: 123
                  message:
                    type: string
                    example: "Marriage license record added successfully"
        '400':
          description: Bad request - missing required fields or invalid file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/licenses/{id}:
    get:
      summary: Get specific marriage license record
      description: Retrieve a single marriage license record by ID
      tags:
        - Marriage Licenses
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Record ID
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Marriage license record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarriageLicense'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete marriage license record
      description: |
        Delete a marriage license record and its associated document file.
        This action cannot be undone.
      tags:
        - Marriage Licenses
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Record ID to delete
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Record deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Record deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/login
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication (username:password)
  schemas:
    MarriageLicense:
      type: object
      properties:
        id:
          type: integer
          description: Unique record identifier
          example: 123
        name_spouse1:
          type: string
          description: First spouse name
          example: "Mary Orlowska"
        name_spouse2:
          type: string
          description: Second spouse name
          example: "Josef PaezKowski"
        marriage_date:
          type: string
          format: date
          description: Marriage date
          example: "1890-02-12"
        license_number:
          type: string
          description: License or application number
          example: "37014"
        file_path:
          type: string
          description: Internal file path
          example: "1759286121101-114686440.pdf"
        original_filename:
          type: string
          description: Original uploaded filename
          example: "marriage_license_1890.pdf"
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
          example: "2024-01-15T10:30:00Z"
      required:
        - id
        - name_spouse1
        - name_spouse2
        - marriage_date
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Record not found"
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
      required:
        - error
...